requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
Jinja2==3.1.2
pymongo==4.8.0
PyYAML==6.0.2
Werkzeug==3.0.4
